Updated the elo-browser to javafx 1.1. Updated the JFXtras library to the latest version.

+++ elo-browser/src/main/java/eu/scy/elobrowser/model/mapping/DisplayMapping.java
@@ -13,5 +13,5 @@ public interface DisplayMapping
 
 	public DisplayProperty getDisplayProperty();
 
-	public Double getValue();
+	public Float getValue();
 }
